fix: unnecessary returns deleted, folders rename and other correction according to review in PR #10
